+++ Classes/Configuration/EventEditorConfigurationCheck.php
@@ -80,12 +80,8 @@ class EventEditorConfigurationCheck extends AbstractFrontEndConfigurationCheck
                 'target_groups',
                 'price_regular',
                 'price_regular_early',
-                // @deprecated #1773 will be removed in seminars 5.0
-                'price_regular_board',
                 'price_special',
                 'price_special_early',
-                // @deprecated #1773 will be removed in seminars 5.0
-                'price_special_board',
                 'payment_methods',
                 'place',
                 'room',
@@ -132,12 +128,8 @@ class EventEditorConfigurationCheck extends AbstractFrontEndConfigurationCheck
                 'target_groups',
                 'price_regular',
                 'price_regular_early',
-                // @deprecated #1773 will be removed in seminars 5.0
-                'price_regular_board',
                 'price_special',
                 'price_special_early',
-                // @deprecated #1773 will be removed in seminars 5.0
-                'price_special_board',
                 'payment_methods',
                 'place',
                 'room',