+++ Configuration/TCA/tx_seminars_seminars.php
@@ -494,23 +494,6 @@ $tca = [
                 'default' => '0.00',
             ],
         ],
-        // @deprecated #1773 will be removed in seminars 5.0
-        'price_regular_board' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.price_regular_board',
-            'displayCond' => 'FIELD:price_on_request:REQ:false',
-            'config' => [
-                'type' => 'input',
-                'size' => 10,
-                'max' => 10,
-                'eval' => 'double2',
-                'range' => [
-                    'upper' => '999999.99',
-                    'lower' => '0.00',
-                ],
-                'default' => '0.00',
-            ],
-        ],
         'price_special' => [
             'exclude' => 1,
             'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.price_special',
@@ -536,23 +519,6 @@ $tca = [
                     'FIELD:price_on_request:REQ:false',
                 ],
             ],
-            'config' => [
-                'type' => 'input',
-                'size' => 10,
-                'max' => 10,
-                'eval' => 'double2',
-                'range' => [
-                    'upper' => '999999.99',
-                    'lower' => '0.00',
-                ],
-                'default' => '0.00',
-            ],
-        ],
-        // @deprecated #1773 will be removed in seminars 5.0
-        'price_special_board' => [
-            'exclude' => 1,
-            'label' => 'LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.price_special_board',
-            'displayCond' => 'FIELD:price_on_request:REQ:false',
             'config' => [
                 'type' => 'input',
                 'size' => 10,
@@ -995,14 +961,14 @@ $tca = [
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelOrganizers, organizers, organizing_partners, event_takes_place_reminder_sent, cancelation_deadline_reminder_sent, ' .
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelAttendees, needs_registration, allows_multiple_registrations, attendees_min, attendees_max, queue_size, offline_attendees, organizers_notified_about_minimum_reached, mute_notification_emails, target_groups, skip_collision_check, date_of_last_registration_digest, registrations, ' .
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelLodging, lodgings, foods, ' .
-                '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelPayment, price_on_request, price_regular, price_regular_early, price_regular_board, price_special, price_special_early, price_special_board, payment_methods, ' .
+                '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelPayment, price_on_request, price_regular, price_regular_early, price_special, price_special_early, payment_methods, ' .
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db:tx_seminars_seminars.divLabelAccess, hidden, starttime, endtime, owner_feuser, vips',
         ],
         \OliverKlee\Seminars\Domain\Model\Event\EventInterface::TYPE_EVENT_TOPIC => [
             'showitem' =>
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.divLabelGeneral, object_type, title, subtitle, image, categories, requirements, dependencies, teaser, description, event_type, credit_points, additional_information, uses_terms_2, notes, attached_files, ' .
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.divLabelAttendees, allows_multiple_registrations, target_groups, ' .
-                '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.divLabelPayment, price_on_request, price_regular, price_regular_early, price_regular_board, price_special, price_special_early, price_special_board, payment_methods, ' .
+                '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.divLabelPayment, price_on_request, price_regular, price_regular_early, price_special, price_special_early, payment_methods, ' .
                 '--div--;LLL:EXT:seminars/Resources/Private/Language/locallang_db.xlf:tx_seminars_seminars.divLabelAccess, hidden, starttime, endtime',
         ],
         \OliverKlee\Seminars\Domain\Model\Event\EventInterface::TYPE_EVENT_DATE => [